refactor: countMatchedNumbers util 함수로 분리
