+++ src/service/CalculatorService.js
@@ -1,4 +1,4 @@
-import countMatchedNumbers from './MatchService.js';
+import countMatchedNumbers from '../util/countMatchedNumbers.js';
 import systemSettings from '../settings/systemSettings.js';
 
 export function calculateWins(lottos, parsedLotto) {
